Skip sphinx build test on Windows

+++ tests/test_docs.py
@@ -55,17 +55,19 @@ class TestDocs:
     def test_build(self) -> None:
         """Try building the docs and see if it works."""
 
-        # Remove the build directory if it exists.
-        if os.path.isdir(os.path.join(self.docs_dir, "build")):
-            shutil.rmtree(os.path.join(self.docs_dir, "build"))
+        # Don't test on Windows
+        if os.name != "nt":
+            # Remove the build directory if it exists.
+            if os.path.isdir(os.path.join(self.docs_dir, "build")):
+                shutil.rmtree(os.path.join(self.docs_dir, "build"))
 
-        return_code = sphinx.cmd.build.main(
-            [
-                "-j",
-                "1",
-                os.path.join(self.docs_dir, "source"),
-                os.path.join(self.docs_dir, "build", "html"),
-            ]
-        )
+            return_code = sphinx.cmd.build.main(
+                [
+                    "-j",
+                    "1",
+                    os.path.join(self.docs_dir, "source"),
+                    os.path.join(self.docs_dir, "build", "html"),
+                ]
+            )
 
-        assert return_code == 0
+            assert return_code == 0
